cleanup: Rename finishedDetourFactory.ts to detourFactory.ts
