+++ assets/tests/components/detours/detourMap.test.tsx
@@ -5,7 +5,7 @@ import { DetourMap } from "../../../src/components/detours/detourMap"
 import "@testing-library/jest-dom/jest-globals"
 import { defaultCenter } from "../../../src/components/map"
 import { latLngLiteralFactory } from "../../factories/latLngLiteralFactory"
-import { routeSegmentsFactory } from "../../factories/finishedDetourFactory"
+import { routeSegmentsFactory } from "../../factories/detourFactory"
 import { RealDispatchWrapper } from "../../testHelpers/wrappers"
 import { layersControlButton } from "../../testHelpers/selectors/components/map"
 import { mockTileUrls } from "../../testHelpers/mockHelpers"
