+++ assets/tests/components/detours/diversionPage.test.tsx
@@ -31,7 +31,7 @@ import {
   finishedDetourFactory,
   routeSegmentsFactory,
   unfinishedDetourFactory,
-} from "../../factories/finishedDetourFactory"
+} from "../../factories/detourFactory"
 import { detourShapeFactory } from "../../factories/detourShapeFactory"
 import {
   missedStopIcon,
